wiregrid_actuator: Move wg actuator agent into package
