+++ socs/agents/wiregrid_actuator/stopper_config.py
